requirements.txt
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Django==2.2.5
djangorestframework==3.10.3
google-api-core==1.14.2
google-auth==1.6.3
google-cloud-vision==0.39.0
googleapis-common-protos==1.6.0
grpcio==1.24.0
idna==2.8
Pillow==6.1.0
protobuf==3.9.2
pyasn1==0.4.7
pyasn1-modules==0.2.6
pytz==2019.2
requests==2.22.0
rsa==4.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.6
